+++ app/src/main/java/org/simple/clinic/monthlyReports/questionnaire/entry/QuestionnaireEntryScreen.kt
@@ -101,7 +101,6 @@ class QuestionnaireEntryScreen : BaseScreen<
     }
   }
 
-
   override fun showProgress() {
   }