+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add Questionnaire screening form json parser
 - Mobius arch setup for questionnaire form
 - Epoxy model setup for questionnaire form
+- Add `QuestionnaireResponse`s table
 
 ## 2023-01-09-8572